refactor(nomic): split dev coordination helper modules
Move lease, receipt, and salvage store logic into dedicated modules so dev_coordination.py stays import-compatible but easier to navigate and maintain.
